Acabado blog

+++ app/Http/Controllers/PostController.php
@@ -80,7 +80,10 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = Post::findOrFail($id);
-        return view('posts.edit', compact('post'));
+        if(auth()->user()->id===$post->usuario_id||auth()->user()->rol==='admin')
+            return view('posts.edit', compact('post'));
+        else
+            return redirect()->route('posts.index');
     }
 
     /**
@@ -92,8 +95,12 @@ class PostController extends Controller
      */
     public function update(PostRequest $request, $id)
     {
-        Post::findOrFail($id)->update($request->all());
+        $post = Post::findOrFail($id);
+        if(auth()->user()->id===$post->usuario_id||auth()->user()->rol==='admin'){
+            $post->update($request->all());
+        }
         return redirect()->route('posts.index');
+
     }
 
     /**
@@ -104,8 +111,11 @@ class PostController extends Controller
      */
     public function destroy($id)
     {
-        Comentario::where('post_id', $id)->delete();
-        Post::findOrFail($id)->delete();
+        $post = Post::findOrFail($id);
+        if(auth()->user()->id===$post->usuario_id||auth()->user()->rol==='admin'){
+            Comentario::where('post_id', $id)->delete();
+            $post->delete();
+        }
         return redirect()->route('posts.index');
     }
